requirements.txt
litellm
sentence-transformers

tqdm
pydantic
loguru
python-dotenv
PyPDF2
python-docx
python-pptx
bs4
markdownify
